requirements.txt
google-cloud-texttospeech~=2.11.1